Add delete review confirmation

+++ src/public/script.js
@@ -6,16 +6,18 @@ document.addEventListener('DOMContentLoaded', function() {
   if (deleteButtons) {
     deleteButtons.forEach(function(button) {
       button.addEventListener('click', function(event) {
-        const reviewID = button.getAttribute('data-id')
-        fetch(`/albums/reviews/${reviewID}`, {
-          method: 'DELETE',
-          credentials: 'same-origin'
-        })
+        if(confirm('Are you sure you want to delete this review?')) {
+          const reviewID = button.getAttribute('data-id')
+          fetch(`/albums/reviews/${reviewID}`, {
+            method: 'DELETE',
+            credentials: 'same-origin'
+          })
           .then(response => {
             console.log(response)
             event.target.parentElement.parentElement.remove()
           })
           .catch(console.error)
+        }
       })
     })
   }
